requirements.txt
av==10.0.0
cffi==1.15.1
confluent-kafka==1.9.2
customtkinter==5.0.3
darkdetect==0.8.0
evdev==1.6.0
kafka-python==2.0.2
MouseInfo==0.1.3
numpy==1.24.0
Pillow==9.3.0
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-xlib==0.32
python3-xlib==0.15
pytweening==1.0.4
six==1.16.0
sounddevice==0.4.5
soundfile==0.11.0
